backend/requirements.txt
# UA2-125 AI Chatbot Assistant - Python Dependencies

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0

# OpenAI API
httpx==0.27.2  # Must be installed BEFORE openai to prevent upgrade to 0.28.x
openai==1.54.0

# Data Processing
numpy==1.26.4
pydantic==2.9.0

# Database
psycopg2-binary==2.9.9
pgvector==0.2.5
sqlalchemy==2.0.23

# Document Processing
PyPDF2==3.0.1  # Fallback for simple PDFs
python-docx==1.1.0
pdf2image==1.16.3  # Convert PDF pages to images
Pillow==10.3.0  # Image processing

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.12

# Optional: For production deployment
gunicorn==23.0.0
